+++ sw/airborne/math/EnuCoor_i.java
@@ -1,7 +1,7 @@
 package sw.airborne.math;
 
 public class EnuCoor_i{
-	public int x, y ,z;
+	public Long x, y ,z;
 	public boolean notequals0(){
 		if(x!=0 && y!= 0 & z!= 0) return true;
 		else return false;
